refactor(authservice): add types to auth service
